pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task_manager - add description"
authors = ["Tim Volkov <volkov.timm@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
uvicorn = "^0.27.1"
psycopg2-binary = "^2.9.9"
django-bootstrap5 = "^23.4"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
rollbar = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
